requirements.txt
#Run in the terminal to install dependencies:
#python -m pip install -r requirements.txt


beautifulsoup4
google-genai
httpx
langchain-text-splitters
python-dotenv
supabase
sentence-transformers
torch
langchain-core
langchain-google-genai
langgraph
numpy
pandas
termcolor
yfinance
